[Feat] KAN-48 상품 교환 API 연결 응답 처리 수정

+++ schoolmate/src/main/java/com/spring/schoolmate/repository/ProductExchangeRepository.java
@@ -10,9 +10,11 @@ import org.springframework.stereotype.Repository;
 public interface ProductExchangeRepository extends JpaRepository<ProductExchange, Integer> {
 
   // 학생 ID로 교환 상품 목록을 페이지네이션하여 조회
-  Page<ProductExchange> findByStudent_StudentIdOrderByExchangeDateDesc(Integer studentId, Pageable pageable);
+  // 🚨 Student ID의 타입 Long으로 변경
+  Page<ProductExchange> findByStudent_StudentIdOrderByExchangeDateDesc(Long studentId, Pageable pageable);
 
   // 사용 상태별로 교환 상품을 페이지네이션하여 조회 (예: '사용중', '사용완료')
+  // 🚨 Student ID의 타입 Long으로 변경
   Page<ProductExchange> findByStudent_StudentIdAndExchangeCardStatusOrderByExchangeDateDesc(
-    Integer studentId, String status, Pageable pageable);
+    Long studentId, String status, Pageable pageable);
 }
